Fitur 2 fckn done, but shit dont know why Fitur 4 jadi rusak

+++ src/main/java/com/apap/tugas1/repository/JabatanPegawaiDB.java
@@ -3,5 +3,7 @@ package com.apap.tugas1.repository;
 import com.apap.tugas1.model.JabatanPegawaiModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JabatanPegawaiDB  extends JpaRepository<JabatanPegawaiModel, Long> {
+import java.math.BigInteger;
+
+public interface JabatanPegawaiDB  extends JpaRepository<JabatanPegawaiModel, BigInteger> {
 }
